refactor code, added new component AddItem
