+++ client/src/components/ItemsList.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import axios from 'axios';
 import NavBar from './NavBar';
-import { Container, ListGroup, ListGroupItem, Input, Form, FormGroup, Button } from 'reactstrap';
+import AddItem from './AddItem';
+import { Container, ListGroup, ListGroupItem, Input, Button } from 'reactstrap';
 
 class ItemsList extends Component {
   constructor(props) {
@@ -11,64 +12,7 @@ class ItemsList extends Component {
       error: null,
       items: [],
       name: ""
-    }
-  }
-  // AJAX call to display all items
-  componentDidMount() {
-    axios.get('/api/items')
-      .then(response => {
-        this.setState({
-          isLoading: true,
-          items: response.data
-        })
-      })    
-  }
-
-  handleValidation = () => {
-    let inputText = this.state.name;
-    let errors = {};
-    let formIsValid = true;
-    //Name
-    if (!inputText) {
-      formIsValid = false;
-      errors["name"] = "Cannot be empty";
-    }
-    if (typeof inputText === "undefined" || inputText === null || !inputText.match(/^\w+( \w+)*$/)) {
-      formIsValid = false;
-      errors["name"] = "Only letters";
-    }
-    this.setState({ errors: errors });
-    return formIsValid;
-  }
-
-  // POST request to add new item
-  onSubmit = (e) => {
-    e.preventDefault();
-    if (!this.handleValidation()) {
-      alert("Error: Please enter letters for your new task, no spaces in the end please");
-      document.getElementById("item").value = "";
-      return
-    }
-    const newItem = {
-      name: this.state.name
-    }
-    axios.post('/api/items', newItem)
-      .then(res => {
-        document.getElementById("item").value = "";
-        this.setState(state => ({
-          items: [res.data, ...state.items], // name: res.data.name 
-        }))
-      })
-  }
-  // Deleting an item
-  deleteItem = (id) => {
-    axios.delete(`/api/items/${id}`)
-      .then(res => {
-        this.setState({
-          // Returning new state by filter method
-          items: this.state.items.filter(item => item._id !== id)
-        })
-      })
+    };
   }
 
   allItems = () => {
@@ -85,6 +29,27 @@ class ItemsList extends Component {
     })
     
   }
+  // AJAX call to display all items
+  componentDidMount() {
+    this.allItems();    
+  }
+
+  addItem = (input) => {
+    this.setState(state => ({
+      items: [input, ...state.items],
+    }))
+  }
+  
+  // Deleting an item
+  deleteItem = (id) => {
+    axios.delete(`/api/items/${id}`)
+      .then(res => {
+        this.setState({
+          // Returning new state by filter method
+          items: this.state.items.filter(item => item._id !== id)
+        })
+      })
+  }  
 
   updateItem = (id, item) => {
     console.log(item);
@@ -106,17 +71,7 @@ class ItemsList extends Component {
           <NavBar items={items}/>       
           <ListGroup>
             <Container>
-              <Form onSubmit={this.onSubmit}>
-                <FormGroup>
-                  <Input required
-                    type="text"
-                    name="name"
-                    id="item"
-                    placeholder="What's next on your mind..."
-                    onChange={(e) => this.setState({ name: e.target.value })}
-                  />
-                </FormGroup>
-              </Form>
+              <AddItem addItem={this.addItem} />              
               <Container>
                 {items.map(item => (
                   <ListGroupItem key={item._id}>
